+++ Misc/NEWS.d/next/Library/2020-05-28-02-08-55.bpo-40799.Q04fYj.rst
@@ -1,0 +1,2 @@
+Move the Python implementation of the :mod:`datetime` module into a new
+``Lib/_pydatetime.py`` file to optimize ``import datetime``.
